E2: Set background colours
Issue: I wanted to make a moving gradient background that will be triggered by a button but it was unsuccessful because I don't know how to make it work with addEventListener when the moving gradient background is in a CSS file.

Nonetheless, the static gradient background works as well. I would prefer to make it move so I will revisit later to find a solution (toggle() doesn't work with my concept as the background is hidden by the other background set in JS file and I want it to appear on click).